[FSTORE-1522] Deduplicate most of hsml (#322)
* Add missing init in hopsworks_common.engine

* Add missing FeatureStoreEncoder to hsfs.util for backwards compatibility

* Deduplicate util

* Deduplicate kafka_topic

* Make direct imports instead of through aliases

* Deduplicate tag

* Deduplicate dataset_api

* Fix tag import

* Fix KafkaTopic

* Fix mocking of get_obj_from_json

* Fix mocking

* Fix mocking of hsml.util

* Fix mocking and imports of util

* Fix KafkaTopic.from_reposnse_json

* Fix pandas typechecking in util

* Fix mocking in test_predictor

* Fix mocking in test_resources

* Fix mocking in test_transformer
