+++ python/hopsworks_common/engine/__init__.py
@@ -1,5 +1,5 @@
 #
-#   Copyright 2021 Logical Clocks AB
+#   Copyright 2024 Hopsworks AB
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -13,12 +13,3 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
-
-from hopsworks_common.tag import (
-    Tag,
-)
-
-
-__all__ = [
-    Tag,
-]
