+++ python/hopsworks_common/kafka_topic.py
@@ -14,19 +14,25 @@
 #   limitations under the License.
 #
 
+from __future__ import annotations
+
 import json
+from typing import Optional
 
 import humps
 from hopsworks_common import usage, util
+from hopsworks_common.constants import KAFKA_TOPIC
 from hopsworks_common.core import kafka_api
 
 
 class KafkaTopic:
+    """Configuration for a Kafka topic."""
+
     def __init__(
         self,
-        name=None,
-        num_of_replicas=None,
-        num_of_partitions=None,
+        name: Optional[str] = KAFKA_TOPIC.CREATE,
+        num_of_replicas: Optional[int] = None,
+        num_of_partitions: Optional[int] = None,
         schema_name=None,
         schema_version=None,
         schema_content=None,
@@ -37,11 +43,18 @@ class KafkaTopic:
         expand=None,
         items=None,
         count=None,
+        num_replicas: Optional[int] = None,
+        num_partitions: Optional[int] = None,
         **kwargs,
     ):
         self._name = name
-        self._num_of_replicas = num_of_replicas
-        self._num_of_partitions = num_of_partitions
+        if not num_of_replicas:
+            num_of_replicas = num_replicas
+        if not num_of_partitions:
+            num_of_partitions = num_partitions
+        self._num_of_replicas, self._num_of_partitions = self._validate_topic_config(
+            self._name, num_of_replicas, num_of_partitions
+        )
         self._schema_name = schema_name
         self._schema_version = schema_version
         self._schema_content = schema_content
@@ -50,15 +63,63 @@ class KafkaTopic:
 
         self._kafka_api = kafka_api.KafkaApi()
 
+    def describe(self):
+        util.pretty_print(self)
+
+    @classmethod
+    def _validate_topic_config(cls, name, num_replicas, num_partitions):
+        if name is not None and name != KAFKA_TOPIC.NONE:
+            if name == KAFKA_TOPIC.CREATE:
+                if num_replicas is None:
+                    print(
+                        "Setting number of replicas to default value '{}'".format(
+                            KAFKA_TOPIC.NUM_REPLICAS
+                        )
+                    )
+                    num_replicas = KAFKA_TOPIC.NUM_REPLICAS
+                if num_partitions is None:
+                    print(
+                        "Setting number of partitions to default value '{}'".format(
+                            KAFKA_TOPIC.NUM_PARTITIONS
+                        )
+                    )
+                    num_partitions = KAFKA_TOPIC.NUM_PARTITIONS
+            else:
+                if num_replicas is not None or num_partitions is not None:
+                    raise ValueError(
+                        "Number of replicas or partitions cannot be changed in existing kafka topics."
+                    )
+        elif name is None or name == KAFKA_TOPIC.NONE:
+            num_replicas = None
+            num_partitions = None
+
+        return num_replicas, num_partitions
+
     @classmethod
     def from_response_json(cls, json_dict):
         json_decamelized = humps.decamelize(json_dict)
         if "count" not in json_decamelized:
-            return cls(**json_decamelized)
+            return cls.from_json(json_decamelized)
         elif json_decamelized["count"] == 0:
             return []
         else:
-            return [cls(**kafka_topic) for kafka_topic in json_decamelized["items"]]
+            return [cls.from_json(jd) for jd in json_decamelized["items"]]
+
+    @classmethod
+    def from_json(cls, json_decamelized):
+        return KafkaTopic(**cls.extract_fields_from_json(json_decamelized))
+
+    @classmethod
+    def extract_fields_from_json(cls, json_decamelized):
+        kwargs = {}
+        kwargs["name"] = json_decamelized.pop("name")  # required
+        kwargs["num_replicas"] = util.extract_field_from_json(
+            json_decamelized, ["num_of_replicas", "num_replicas"]
+        )
+        kwargs["num_partitions"] = util.extract_field_from_json(
+            json_decamelized, ["num_of_partitions", "num_partitions"]
+        )
+        return kwargs
 
     def update_from_response_json(self, json_dict):
         json_decamelized = humps.decamelize(json_dict)
@@ -67,18 +128,40 @@ class KafkaTopic:
 
     @property
     def name(self):
-        """Name of the topic"""
+        """Name of the Kafka topic."""
         return self._name
+
+    @name.setter
+    def name(self, name: str):
+        self._name = name
 
     @property
     def replicas(self):
-        """Replication factor for the topic"""
+        """Number of replicas of the Kafka topic."""
         return self._num_of_replicas
 
     @property
+    def num_replicas(self):
+        """Number of replicas of the Kafka topic."""
+        return self._num_of_replicas
+
+    @num_replicas.setter
+    def num_replicas(self, num_replicas: int):
+        self._num_of_replicas = num_replicas
+
+    @property
     def partitions(self):
-        """Number of partitions for the topic"""
+        """Number of partitions of the Kafka topic."""
         return self._num_of_partitions
+
+    @property
+    def num_partitions(self):
+        """Number of partitions of the Kafka topic."""
+        return self._num_of_partitions
+
+    @num_partitions.setter
+    def topic_num_partitions(self, num_partitions: int):
+        self._num_partitions = num_partitions
 
     @property
     def schema(self):
@@ -99,6 +182,15 @@ class KafkaTopic:
 
     def json(self):
         return json.dumps(self, cls=util.Encoder)
+
+    def to_dict(self):
+        return {
+            "kafkaTopicDTO": {
+                "name": self._name,
+                "numOfReplicas": self._num_of_replicas,
+                "numOfPartitions": self._num_of_partitions,
+            }
+        }
 
     def __str__(self):
         return self.json()
