+++ python/hsml/model.py
@@ -21,10 +21,9 @@ import warnings
 from typing import Any, Dict, Optional, Union
 
 import humps
-from hopsworks_common import usage
-from hsml import client, util
-from hsml.constants import ARTIFACT_VERSION
-from hsml.constants import INFERENCE_ENDPOINTS as IE
+from hopsworks_common import client, usage, util
+from hopsworks_common.constants import ARTIFACT_VERSION
+from hopsworks_common.constants import INFERENCE_ENDPOINTS as IE
 from hsml.core import explicit_provenance
 from hsml.engine import model_engine
 from hsml.inference_batcher import InferenceBatcher
@@ -360,7 +359,7 @@ class Model:
         return self
 
     def json(self):
-        return json.dumps(self, cls=util.MLEncoder)
+        return json.dumps(self, cls=util.Encoder)
 
     def to_dict(self):
         return {
