+++ python/tests/test_deployable_component.py
@@ -40,7 +40,7 @@ class TestDeployableComponent:
     def test_constructor_default(self, mocker):
         # Arrange
         mock_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=None
+            "hopsworks_common.util.get_obj_from_json", return_value=None
         )
         mock_ib_init = mocker.patch(
             "hsml.inference_batcher.InferenceBatcher.__init__", return_value=None
@@ -73,7 +73,7 @@ class TestDeployableComponent:
         resources = {}
         inf_batcher = inference_batcher.InferenceBatcher()
         mock_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=inf_batcher
+            "hopsworks_common.util.get_obj_from_json", return_value=inf_batcher
         )
         mock_ib_init = mocker.patch(
             "hsml.inference_batcher.InferenceBatcher.__init__", return_value=None
