+++ python/tests/test_deployment.py
@@ -634,7 +634,7 @@ class TestDeployment:
         p = self._get_dummy_predictor(mocker, backend_fixtures)
         d = deployment.Deployment(predictor=p)
         mock_util_get_members = mocker.patch(
-            "hsml.util.get_members", return_value=["predictor"]
+            "hopsworks_common.util.get_members", return_value=["predictor"]
         )
         mock_print = mocker.patch("builtins.print")
 
@@ -661,7 +661,7 @@ class TestDeployment:
         p = self._get_dummy_predictor(mocker, backend_fixtures)
         d = deployment.Deployment(predictor=p)
         mock_util_get_members = mocker.patch(
-            "hsml.util.get_members", return_value=["valid"]
+            "hopsworks_common.util.get_members", return_value=["valid"]
         )
         mock_print = mocker.patch("builtins.print")
 
@@ -700,7 +700,7 @@ class TestDeployment:
         p = self._get_dummy_predictor(mocker, backend_fixtures)
         d = deployment.Deployment(predictor=p)
         mock_util_get_members = mocker.patch(
-            "hsml.util.get_members", return_value=["predictor"]
+            "hopsworks_common.util.get_members", return_value=["predictor"]
         )
         mock_print = mocker.patch("builtins.print")
 
@@ -727,7 +727,7 @@ class TestDeployment:
         p = self._get_dummy_predictor(mocker, backend_fixtures)
         d = deployment.Deployment(predictor=p)
         mock_util_get_members = mocker.patch(
-            "hsml.util.get_members", return_value=["predictor"]
+            "hopsworks_common.util.get_members", return_value=["predictor"]
         )
         mock_print = mocker.patch("builtins.print")
 
@@ -758,10 +758,10 @@ class TestDeployment:
         path = "/p/" + str(mock_client._project_id) + "/deployments/" + str(d.id)
 
         mock_util_get_hostname_replaced_url = mocker.patch(
-            "hsml.util.get_hostname_replaced_url", return_value="url"
+            "hopsworks_common.util.get_hostname_replaced_url", return_value="url"
         )
         mock_client_get_instance = mocker.patch(
-            "hsml.client.get_instance", return_value=mock_client
+            "hopsworks_common.client.get_instance", return_value=mock_client
         )
 
         # Act
@@ -781,7 +781,7 @@ class TestDeployment:
         mocker.patch("hsml.predictor.Predictor._validate_serving_tool")
         mocker.patch("hsml.predictor.Predictor._validate_resources")
         mocker.patch("hsml.predictor.Predictor._validate_script_file")
-        mocker.patch("hsml.util.get_obj_from_json")
+        mocker.patch("hopsworks_common.util.get_obj_from_json")
         return predictor.Predictor(
             id=p_json["id"],
             name=p_json["name"],
