+++ python/tests/test_inference_logger.py
@@ -107,7 +107,7 @@ class TestInferenceLogger:
         )
         default_kt = kafka_topic.KafkaTopic()
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=default_kt
+            "hopsworks_common.util.get_obj_from_json", return_value=default_kt
         )
 
         # Act
@@ -136,7 +136,7 @@ class TestInferenceLogger:
         )
         default_kt = kafka_topic.KafkaTopic()
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=default_kt
+            "hopsworks_common.util.get_obj_from_json", return_value=default_kt
         )
 
         # Act
@@ -163,7 +163,7 @@ class TestInferenceLogger:
         )
         default_kt = kafka_topic.KafkaTopic()
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=default_kt
+            "hopsworks_common.util.get_obj_from_json", return_value=default_kt
         )
 
         # Act
@@ -190,7 +190,7 @@ class TestInferenceLogger:
         )
         default_kt = kafka_topic.KafkaTopic()
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=default_kt
+            "hopsworks_common.util.get_obj_from_json", return_value=default_kt
         )
 
         # Act
@@ -220,7 +220,7 @@ class TestInferenceLogger:
         )
         kt = kafka_topic.KafkaTopic(json["kafka_topic"]["name"])
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=kt
+            "hopsworks_common.util.get_obj_from_json", return_value=kt
         )
 
         # Act
@@ -250,7 +250,7 @@ class TestInferenceLogger:
         )
         kt = kafka_topic.KafkaTopic(json["kafka_topic"]["name"])
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=kt
+            "hopsworks_common.util.get_obj_from_json", return_value=kt
         )
 
         # Act
@@ -280,7 +280,7 @@ class TestInferenceLogger:
         )
         kt = kafka_topic.KafkaTopic(json["kafka_topic"]["name"])
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=kt
+            "hopsworks_common.util.get_obj_from_json", return_value=kt
         )
 
         # Act
@@ -310,7 +310,7 @@ class TestInferenceLogger:
         )
         kt = kafka_topic.KafkaTopic(json["kafka_topic"]["name"])
         mock_util_get_obj_from_json = mocker.patch(
-            "hsml.util.get_obj_from_json", return_value=kt
+            "hopsworks_common.util.get_obj_from_json", return_value=kt
         )
 
         # Act
