+++ python/tests/test_kafka_topic.py
@@ -18,8 +18,8 @@ import copy
 
 import humps
 import pytest
-from hsml import kafka_topic
-from hsml.constants import KAFKA_TOPIC
+from hopsworks_common import kafka_topic
+from hopsworks_common.constants import KAFKA_TOPIC
 
 
 class TestKafkaTopic:
@@ -31,7 +31,9 @@ class TestKafkaTopic:
             "response"
         ]["kafka_topic_dto"]
         json_camelized = humps.camelize(json)  # as returned by the backend
-        mock_kt_from_json = mocker.patch("hsml.kafka_topic.KafkaTopic.from_json")
+        mock_kt_from_json = mocker.patch(
+            "hopsworks_common.kafka_topic.KafkaTopic.from_json"
+        )
 
         # Act
         _ = kafka_topic.KafkaTopic.from_response_json(json_camelized)
@@ -45,7 +47,9 @@ class TestKafkaTopic:
             "response"
         ]["kafka_topic_dto"]
         json_camelized = humps.camelize(json)  # as returned by the backend
-        mock_kt_from_json = mocker.patch("hsml.kafka_topic.KafkaTopic.from_json")
+        mock_kt_from_json = mocker.patch(
+            "hopsworks_common.kafka_topic.KafkaTopic.from_json"
+        )
 
         # Act
         _ = kafka_topic.KafkaTopic.from_response_json(json_camelized)
@@ -61,10 +65,11 @@ class TestKafkaTopic:
             "response"
         ]["kafka_topic_dto"]
         mock_kt_extract_fields = mocker.patch(
-            "hsml.kafka_topic.KafkaTopic.extract_fields_from_json", return_value=json
+            "hopsworks_common.kafka_topic.KafkaTopic.extract_fields_from_json",
+            return_value=json,
         )
         mock_kt_init = mocker.patch(
-            "hsml.kafka_topic.KafkaTopic.__init__", return_value=None
+            "hopsworks_common.kafka_topic.KafkaTopic.__init__", return_value=None
         )
 
         # Act
@@ -80,10 +85,11 @@ class TestKafkaTopic:
             "response"
         ]["kafka_topic_dto"]
         mock_kt_extract_fields = mocker.patch(
-            "hsml.kafka_topic.KafkaTopic.extract_fields_from_json", return_value=json
+            "hopsworks_common.kafka_topic.KafkaTopic.extract_fields_from_json",
+            return_value=json,
         )
         mock_kt_init = mocker.patch(
-            "hsml.kafka_topic.KafkaTopic.__init__", return_value=None
+            "hopsworks_common.kafka_topic.KafkaTopic.__init__", return_value=None
         )
 
         # Act
@@ -101,7 +107,7 @@ class TestKafkaTopic:
             "response"
         ]["kafka_topic_dto"]
         mock_kt_validate_topic_config = mocker.patch(
-            "hsml.kafka_topic.KafkaTopic._validate_topic_config",
+            "hopsworks_common.kafka_topic.KafkaTopic._validate_topic_config",
             return_value=(KAFKA_TOPIC.NUM_REPLICAS, KAFKA_TOPIC.NUM_PARTITIONS),
         )
 
@@ -121,7 +127,7 @@ class TestKafkaTopic:
             "response"
         ]["kafka_topic_dto"]
         mock_kt_validate_topic_config = mocker.patch(
-            "hsml.kafka_topic.KafkaTopic._validate_topic_config",
+            "hopsworks_common.kafka_topic.KafkaTopic._validate_topic_config",
             return_value=(json["num_replicas"], json["num_partitions"]),
         )
